Added calls to regular checks to the controllers and refactoring out some no more needed functions
