+++ lib/Controller/ConfigController.php
@@ -46,9 +46,11 @@ class ConfigController extends Controller
      * @NoCSRFRequired
      */
     public function list() {
+        $this->dbCacheService->triggerCheck();
+        
         return new DataResponse([
             'folder' => $this->service->getUserFolderPath(),
-            'update_interval' => $this->service->getSearchIndexUpdateInterval(),
+            'update_interval' => $this->dbCacheService->getSearchIndexUpdateInterval(),
             'print_image' => $this->service->getPrintImage(),
         ], Http::STATUS_OK);
     }
@@ -59,9 +61,11 @@ class ConfigController extends Controller
      */
     public function config()
     {
+        $this->dbCacheService->triggerCheck();
+        
         if (isset($_POST['folder'])) {
             $this->service->setUserFolderPath($_POST['folder']);
-            $this->service->rebuildSearchIndex();
+            $this->dbCacheService->updateCache();
         }
 
         if (isset($_POST['update_interval'])) {
@@ -81,7 +85,6 @@ class ConfigController extends Controller
      */
     public function reindex()
     {
-        //$this->service->rebuildSearchIndex();
         $this->dbCacheService->updateCache();
 
         return new DataResponse('Search index rebuilt successfully', Http::STATUS_OK);
